ApiExceptionHandler.java: - Added detail to note that ApiRequestException is equivalent to EntityNotFoundException. - Added new MethodArgumentNotValidException handler. This is triggered when an object fails @Valid validation in Controller.
pom.xml:
- Added spring-boot-starter-validation dependency because @NotEmpty annotation in TeamDTO was not working.

TeamDTO.java:
- Removed @NotEmpty from manager field, as a Team does not technically require a manager in this project.

TeamService.java:
- Removed logic that checked the validity of [PUT] updateTeamById incoming TeamDTO.
- Was able to remove this as @Valid now used in RequestBody of PUT call in controller.

TeamController.java:
- Added @Valid in PUT request handler for incoming TeamDTO.

NOTE: The combined effect of removing the Service Layer "checks" and adding Controller Layer @Valid annotation now means that future PUT requests for Team resource requires:

1. ALL FIELDS in JSON Body

More specifically
1.1 @NotEmpty fields are MANDATORY and if excluded will trigger an exception (due to @Valid)
1.2 Any non-mandatory/optional fields omitted from the JSON Body will be NULLED/REMOVED from the resource you are updating (due to no 'checks')

It was convenient to use PUT request to partially update records HOWEVER, with the importance of validation and correctly abiding by RESTful architecture... the body of the PUT request should replace the entire resource, therefore, this change is correct.

If partial updates are desired/required, we must use the PATCH HTTP request.

"It's the responsibility of the client to create a message that describes what the client actually wants. So if you, the client, intend that the optional fields are unchanged after your request, you need to include them in the representation that you include in the body of the request." - VoiceOfUnreason, StackOverflow